[refacto] renamed format module to file_format
also removed statistics from the PartitionedFile abstraction
