+++ datafusion/src/physical_plan/mod.rs
@@ -620,8 +620,8 @@ pub mod distinct_expressions;
 pub mod empty;
 pub mod explain;
 pub mod expressions;
+pub mod file_format;
 pub mod filter;
-pub mod format;
 pub mod functions;
 pub mod hash_aggregate;
 pub mod hash_join;
